refactor: implementation of RBAC
Role Based Acces control were implemented in the places where it's needed.
