+++ src/common/application/decorators/index.ts
@@ -1,2 +1,2 @@
 export * from './public.decorator';
-export * from './role.decorator';
+export * from './roles.decorator';
