--- src/common/application/decorators/role.decorator.ts
@@ -1,6 +1,0 @@
-import { SetMetadata } from "@nestjs/common";
-
-import { Role } from "../../domain/enums";
-
-export const ROLE_KEY = 'role';
-export const Roles = (...roles: Role[]) => (SetMetadata(ROLE_KEY, roles));
