+++ src/common/domain/enums/index.ts
@@ -1,1 +1,1 @@
-export * from './role.enum';
+export * from './roles.enum';
