+++ src/common/domain/enums/roles.enum.ts
@@ -1,4 +1,4 @@
-export enum Role {
+export enum Roles {
     ADMIN = 'ADMIN',
     CUSTOMER = 'CUSTOMER'
 };
